Deepen spectacular adapter and prune dead URN parsing
